# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dc0d95e9ef7a1181f8e1322
  title: khachsan
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /costomer/:
    post:
      x-skaffolder-id: 5dc0efcd9ef7a1181f8e1481
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc0ea4b9ef7a1181f8e1409
      x-skaffolder-resource: costomer
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - costomer
      parameters:
        - name: obj
          x-skaffolder-type: costomer
          in: path
          description: Object to insert
          required: true
          schema:
            type: costomer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/costomer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                age:
                  type: number
                card:
                  type: number
                coderoom:
                  type: integer
                gender:
                  type: string
                namcostomer:
                  type: string
                nationality:
                  type: string
                phonenumber:
                  type: number
    get:
      x-skaffolder-id: 5dc0efc19ef7a1181f8e147e
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc0ea4b9ef7a1181f8e1409
      x-skaffolder-resource: costomer
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF costomer
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - costomer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/costomer"
  /costomer/{id}:
    delete:
      x-skaffolder-id: 5dc0efd69ef7a1181f8e1483
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc0ea4b9ef7a1181f8e1409
      x-skaffolder-resource: costomer
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - costomer
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id costomer
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc0efe69ef7a1181f8e1486
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0ea4b9ef7a1181f8e1409
      x-skaffolder-resource: costomer
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: costomer
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - costomer
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id costomer
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/costomer"
    post:
      x-skaffolder-id: 5dc0efc79ef7a1181f8e147f
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc0ea4b9ef7a1181f8e1409
      x-skaffolder-resource: costomer
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: costomer
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - costomer
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id costomer
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/costomer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/costomer"
  /costomer/findBynamcostomer/{key}:
    get:
      x-skaffolder-id: 5dc0ef669ef7a1181f8e147b
      x-skaffolder-name: findBynamcostomer
      x-skaffolder-id-resource: 5dc0ea4b9ef7a1181f8e1409
      x-skaffolder-resource: costomer
      x-skaffolder-crudAction: findBynamcostomer
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBynamcostomer
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBynamcostomer/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBynamcostomer
      tags:
        - costomer
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of the resource namcostomer to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/costomer"
    
  #-----  employess  -----
  /employess/:
    post:
      x-skaffolder-id: 5dc0e6759ef7a1181f8e13de
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc0e3dd9ef7a1181f8e13d0
      x-skaffolder-resource: employess
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - employess
      parameters:
        - name: obj
          x-skaffolder-type: employess
          in: path
          description: Object to insert
          required: true
          schema:
            type: employess
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/employess"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                disc:
                  type: string
                gender:
                  type: string
                name:
                  type: string
                position:
                  type: string
                salary:
                  type: number
    get:
      x-skaffolder-id: 5dc0e6579ef7a1181f8e13d9
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc0e3dd9ef7a1181f8e13d0
      x-skaffolder-resource: employess
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF employess
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - employess
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/employess"
  /employess/{id}:
    delete:
      x-skaffolder-id: 5dc0e67e9ef7a1181f8e13e0
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc0e3dd9ef7a1181f8e13d0
      x-skaffolder-resource: employess
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - employess
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id employess
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc0e66c9ef7a1181f8e13dc
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0e3dd9ef7a1181f8e13d0
      x-skaffolder-resource: employess
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: employess
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - employess
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id employess
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/employess"
    post:
      x-skaffolder-id: 5dc0e6649ef7a1181f8e13da
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc0e3dd9ef7a1181f8e13d0
      x-skaffolder-resource: employess
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: employess
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - employess
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id employess
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/employess"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/employess"
  /employess/findByname/{key}:
    get:
      x-skaffolder-id: 5dc0e64c9ef7a1181f8e13d7
      x-skaffolder-name: findByname
      x-skaffolder-id-resource: 5dc0e3dd9ef7a1181f8e13d0
      x-skaffolder-resource: employess
      x-skaffolder-crudAction: findByname
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByname
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByname/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByname
      tags:
        - employess
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of the resource name to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/employess"
    
  #-----  room  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    costomer:
      x-skaffolder-id: 5dc0ea4b9ef7a1181f8e1409
      x-skaffolder-id-db: 5dc0d95e9ef7a1181f8e1323
      x-skaffolder-id-entity: 5dc0ea4b9ef7a1181f8e1401
      x-skaffolder-url: /costomer
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        age:
          type: number
          x-skaffolder-id-attr: 5dc0ea4b9ef7a1181f8e1406
          x-skaffolder-type: Number
        card:
          type: number
          x-skaffolder-id-attr: 5dc0ea4b9ef7a1181f8e1403
          x-skaffolder-type: Number
        coderoom:
          type: integer
          x-skaffolder-id-attr: 5dc0ea4b9ef7a1181f8e1408
          x-skaffolder-type: Integer
        gender:
          type: string
          x-skaffolder-id-attr: 5dc0ea4b9ef7a1181f8e1405
          x-skaffolder-type: String
        namcostomer:
          type: string
          x-skaffolder-id-attr: 5dc0ea4b9ef7a1181f8e1402
          x-skaffolder-type: String
        nationality:
          type: string
          x-skaffolder-id-attr: 5dc0ea4b9ef7a1181f8e1404
          x-skaffolder-type: String
        phonenumber:
          type: number
          x-skaffolder-id-attr: 5dc0ea4b9ef7a1181f8e1407
          x-skaffolder-type: Number
      x-skaffolder-relations: 
      required:
        - _id
    employess:
      x-skaffolder-id: 5dc0e3dd9ef7a1181f8e13d0
      x-skaffolder-id-db: 5dc0d95e9ef7a1181f8e1323
      x-skaffolder-id-entity: 5dc0e3dd9ef7a1181f8e13cf
      x-skaffolder-url: /employess
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        disc:
          type: string
          x-skaffolder-id-attr: 5dc0e6939ef7a1181f8e13e6
          x-skaffolder-type: String
        gender:
          type: string
          x-skaffolder-id-attr: 5dc0e6939ef7a1181f8e13e5
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dc0e6939ef7a1181f8e13e2
          x-skaffolder-type: String
        position:
          type: string
          x-skaffolder-id-attr: 5dc0e6939ef7a1181f8e13e3
          x-skaffolder-type: String
        salary:
          type: number
          x-skaffolder-id-attr: 5dc0e6939ef7a1181f8e13e4
          x-skaffolder-type: Decimal
      x-skaffolder-relations: 
        employessid:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dc0e3dd9ef7a1181f8e13cf
          x-skaffolder-ent2: 5dc0e8119ef7a1181f8e13eb
        employessid:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dc0e3dd9ef7a1181f8e13cf
          x-skaffolder-ent2: 5dc0e8d59ef7a1181f8e13f6
      required:
        - _id
...